semana06/dia5/mongoaatlas/requirements.txt
flask
pymongo
pymongo[srv]
python-dotenv
Flask-PyMongo